refactor: detail url nested로 수정
